modified controller files - added comments - changed localhost - all …

+++ build/controllers/commentController.js
@@ -8,6 +8,7 @@ const getAllComments = async (req, res, next) => {
     res.status(200).json(comments);
 };
 exports.getAllComments = getAllComments;
+// Aaron: we might need to somehow bring in the goalId without making the user enter it.
 const createComment = async (req, res, next) => {
     let user = await (0, auth_1.verifyUser)(req); // user authentication
     if (!user) {
@@ -15,8 +16,10 @@ const createComment = async (req, res, next) => {
     }
     let newComment = req.body;
     newComment.username = user.username;
+    // checking for a comment and a goal 
     if (newComment.comment && newComment.goalId) {
-        let created = await comment_1.Comment.create({ ...newComment, username: user.username, goalId: newComment.goalId });
+        // creating newComment and setting goalId to newComment.goalId
+        let created = await comment_1.Comment.create({ ...newComment, goalId: newComment.goalId });
         await created.save();
         res.status(201).json(created);
     }
@@ -28,6 +31,7 @@ exports.createComment = createComment;
 const getComment = async (req, res, next) => {
     let commentId = req.params.id;
     let commentFound = await comment_1.Comment.findByPk(commentId);
+    //checking for comment
     if (commentFound) {
         res.status(200).json(commentFound);
     }
@@ -43,11 +47,11 @@ const updateComment = async (req, res, next) => {
     }
     let commentId = req.params.id;
     let newComment = req.body;
+    // setting comment username to current user
     newComment.username = user.username;
     newComment.goalId = req.body.goalId;
-    console.log("req.params.id", req.params.id);
-    console.log("req.body", req.body);
     let commentFound = await comment_1.Comment.findByPk(commentId);
+    // checking for a comment, the comment found matches comment being updated, goal id of comment matches new comment goal id, and username from old comment matches username of current user.
     if (commentFound && commentFound.commentId == newComment.commentId
         && commentFound.goalId == newComment.goalId
         && commentFound.username == user.username) {
@@ -68,7 +72,8 @@ const deleteComment = async (req, res, next) => {
     }
     const commentId = req.params.id;
     const commentFound = await comment_1.Comment.findByPk(commentId);
-    if (commentFound) {
+    // checking for comment and if comment's username matches current user.
+    if (commentFound && commentFound.username == user.username) {
         await comment_1.Comment.destroy({
             where: { commentId: commentId }
         });
