+++ src/controllers/userController.ts
@@ -9,10 +9,15 @@ export const getAllUsers: RequestHandler = async (req, res, next) => {
 
 export const createUser: RequestHandler = async (req, res, next) => {
     let newUser: User = req.body;
+
+    // checking for all required fields filled out
     if (newUser.username && newUser.password && newUser.firstName && newUser.email) {
         let hashedPassword = await hashPassword(newUser.password);
+
         newUser.password = hashedPassword;
+
         let created = await User.create(newUser);
+
         res.status(200).json({
             username: created.username,
             userId: created.userId
@@ -25,7 +30,10 @@ export const createUser: RequestHandler = async (req, res, next) => {
 
 export const getUser: RequestHandler = async (req, res, next) => {
     let username = req.params.username;
+
     let user = await User.findByPk(username);
+
+    // checking for a user
     if (user) {
         res.status(200).json(user);
     }
@@ -35,7 +43,7 @@ export const getUser: RequestHandler = async (req, res, next) => {
 }
 
 export const updateUser: RequestHandler = async (req, res, next) => {
-    let user: User | null = await verifyUser(req);
+    let user: User | null = await verifyUser(req); // user authentication
 
     if (!user) {
         return res.status(403).send();
@@ -47,6 +55,7 @@ export const updateUser: RequestHandler = async (req, res, next) => {
 
     let userFound = await User.findByPk(username);
 
+    // checking for found user and username found matches current user
     if (userFound && userFound.username == newUser.username) {
         await User.update(newUser, {
             where: { username: username }
@@ -61,7 +70,7 @@ export const updateUser: RequestHandler = async (req, res, next) => {
 }
 
 export const deleteUser: RequestHandler = async (req, res, next) => {
-    let user: User | null = await verifyUser(req);
+    let user: User | null = await verifyUser(req); // user authentication
 
     if (!user) {
         return res.status(403).send();
@@ -70,7 +79,8 @@ export const deleteUser: RequestHandler = async (req, res, next) => {
     let username = req.params.username;
     let userFound = await User.findByPk(username);
 
-    if (userFound) {
+    // checking if user was found and matches current user
+    if (userFound && userFound.username == user.username) {
         await User.destroy({
             where: { username: username }
         });
